requirements.txt
streamlit 
groq
python-dotenv
langchain
langchain_groq
